requirements.txt
Pillow>=9.0
numpy>=1.22
tqdm>=4.64
fargv>=0.9
torch>=2.0
torchvision>=0.15
rarfile>=4.0
py7zr>=0.20
gdown>=4.7
pytest>=7.0
pytest-cov>=4.0
sphinx>=6.0
myst-parser>=1.0
sphinx-rtd-theme>=1.2
ruff>=0.1
coverage>=7.0
